split code to allow testability
